fix more warnings related with int/float interactions

+++ Aracnoids/src/entities/mouse.cpp
@@ -4,8 +4,8 @@ namespace mouse
 {
 	Mouse CreateMouse(Mouse gameMouse)
 	{
-		gameMouse.mousePos.x =(float) GetMouseX();
-		gameMouse.mousePos.y = GetMouseY();
+		gameMouse.mousePos.x = static_cast<float>(GetMouseX());
+		gameMouse.mousePos.y = static_cast<float>(GetMouseY());
 
 		return gameMouse;
 	}
